requirements.txt
alembic==1.5.8
arrow==1.0.3
bcrypt==3.2.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.4
MarkupSafe==1.1.1
newsapi-python==0.2.6
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.3
